backend/requirements.txt
Django>=6.0,<6.1
psycopg2-binary>=2.9,<3.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.4,<5.0
reportlab>=4.2,<5.0
Pillow>=10.4,<12.0
